+++ benchmarks/backend_throughput/quill_backend_throughput.cpp
@@ -37,8 +37,9 @@ int main()
 
   quill::Logger* logger = quill::Frontend::create_or_get_logger(
     "bench_logger", std::move(file_sink),
-    "%(time) [%(thread_id)] %(short_source_location) %(log_level) %(message)", "%H:%M:%S.%Qns",
-    quill::Timezone::LocalTime, false);
+    quill::PatternFormatterOptions{
+      "%(time) [%(thread_id)] %(short_source_location) %(log_level) %(message)", "%H:%M:%S.%Qns",
+      quill::Timezone::LocalTime, false});
 
   quill::Frontend::preallocate();