+++ examples/advanced/quill_wrapper/include/quill_wrapper/quill_wrapper.cpp
@@ -14,8 +14,9 @@ void setup_quill(char const* log_file)
   auto console_sink = quill::Frontend::create_or_get_sink<quill::ConsoleSink>("console_sink");
 
   // Create and store the logger
-  quill::Frontend::create_or_get_logger("root", std::move(console_sink),
-                                        "%(time) [%(thread_id)] %(short_source_location:<28) "
+  quill::Frontend::create_or_get_logger(
+    "root", std::move(console_sink),
+    quill::PatternFormatterOptions{"%(time) [%(thread_id)] %(short_source_location:<28) "
                                         "LOG_%(log_level:<9) %(logger:<12) %(message)",
-                                        "%H:%M:%S.%Qns", quill::Timezone::GmtTime);
+                                   "%H:%M:%S.%Qns", quill::Timezone::GmtTime});
 }