+++ examples/rotating_file_logging.cpp
@@ -31,11 +31,11 @@ int main()
       return cfg;
     }());
 
-  quill::Logger* logger =
-    quill::Frontend::create_or_get_logger("root", std::move(rotating_file_sink),
-                                          "%(time) [%(thread_id)] %(short_source_location:<28) "
+  quill::Logger* logger = quill::Frontend::create_or_get_logger(
+    "root", std::move(rotating_file_sink),
+    quill::PatternFormatterOptions{"%(time) [%(thread_id)] %(short_source_location:<28) "
                                           "LOG_%(log_level:<9) %(logger:<12) %(message)",
-                                          "%H:%M:%S.%Qns", quill::Timezone::GmtTime);
+                                   "%H:%M:%S.%Qns", quill::Timezone::GmtTime});
 
   for (int i = 0; i < 20; ++i)
   {