+++ examples/shared_library/quill_wrapper_shared/include/quill_wrapper_shared/quill_wrapper_shared.cpp
@@ -17,9 +17,9 @@ void setup_quill()
   auto console_sink = quill::Frontend::create_or_get_sink<quill::ConsoleSink>("sink_id_1");
 
   // Create and store the logger
-  global_logger_a =
-    quill::Frontend::create_or_get_logger("root", std::move(console_sink),
-                                          "%(time) [%(thread_id)] %(short_source_location:<28) "
+  global_logger_a = quill::Frontend::create_or_get_logger(
+    "root", std::move(console_sink),
+    quill::PatternFormatterOptions{"%(time) [%(thread_id)] %(short_source_location:<28) "
                                           "LOG_%(log_level:<9) %(logger:<12) %(message)",
-                                          "%H:%M:%S.%Qns", quill::Timezone::GmtTime);
+                                   "%H:%M:%S.%Qns", quill::Timezone::GmtTime});
 }