+++ examples/system_clock_logging.cpp
@@ -22,8 +22,10 @@ int main()
   auto console_sink = quill::Frontend::create_or_get_sink<quill::ConsoleSink>("sink_id_1");
 
   quill::Logger* logger = quill::Frontend::create_or_get_logger(
-    "root", std::move(console_sink), "%(time) [%(process_id)] [%(thread_id)] %(logger) - %(message)",
-    "%D %H:%M:%S.%Qms %z", quill::Timezone::GmtTime, false, quill::ClockSourceType::System);
+    "root", std::move(console_sink),
+    quill::PatternFormatterOptions{"%(time) [%(process_id)] [%(thread_id)] %(logger) - %(message)",
+                                   "%D %H:%M:%S.%Qms %z", quill::Timezone::GmtTime, false},
+    quill::ClockSourceType::System);
 
   // Change the LogLevel to print everything
   logger->set_log_level(quill::LogLevel::TraceL3);