+++ examples/tags_logging.cpp
@@ -28,9 +28,10 @@ int main()
 
   quill::Logger* logger = quill::Frontend::create_or_get_logger(
     "root", std::move(console_sink),
-    "%(time) [%(thread_id)] %(short_source_location:<28) %(log_level:<9) "
-    "%(tags)%(message)",
-    "%Y-%m-%d %H:%M:%S.%Qms", quill::Timezone::GmtTime);
+    quill::PatternFormatterOptions{
+      "%(time) [%(thread_id)] %(short_source_location:<28) %(log_level:<9) "
+      "%(tags)%(message)",
+      "%Y-%m-%d %H:%M:%S.%Qms", quill::Timezone::GmtTime});
 
   LOG_INFO_TAGS(logger, TAGS("random"), "Debug with tags");
   LOG_INFO_TAGS(logger, TAGS(TAG_2), "Info with tags");