+++ test/integration_tests/JsonFileLoggingTest.cpp
@@ -89,8 +89,9 @@ TEST_CASE("json_file_logging")
         Logger* logger = Frontend::create_or_get_logger(
           logger_name_prefix + std::to_string(i),
           std::initializer_list<std::shared_ptr<Sink>>{std::move(json_file_sink), std::move(file_sink)},
-          "%(time) [%(thread_id)] %(short_source_location:<28) LOG_%(log_level:<9) %(logger:<12) "
-          "%(message) [%(named_args)]");
+          quill::PatternFormatterOptions{
+            "%(time) [%(thread_id)] %(short_source_location:<28) LOG_%(log_level:<9) %(logger:<12) "
+            "%(message) [%(named_args)]"});
 
         if (i == 0)
         {